requirements.txt
svgwrite==1.4.3
pyyaml==6.0.1
